Code shuffling and expansion of tracker, accelerator config args to beyond wandb
